Set AWS Lambda reserved concurrency (max), queue-concurrency (starting workers), and cli-concurrency (required for cli scheduler).
This commit sets a total of 41 concurrent Lambda functions. Will fail after reaching the AWS default limit of 1000 concurrent functions. This can be raised by contacting AWS and asking to have that limit increased if needed.
Reserved concurrency containters are free, not only reserves concurrency slots, but will also limit concurrency. Provisioned concurrency is not free but can be used to avoid this and will be used in a future commit.